backend/subscription_service/requirements.txt
google-cloud-firestore
google-cloud-secret-manager
firebase-admin
functions-framework
Flask
razorpay